Added asking the user to migrate and using the curr-symlink

+++ tor-browser/check_and_install.sh
@@ -49,12 +49,16 @@ ln -s tor-browser* curr || exit 1
 
 
 # migrate directories in the TRANSITIONS-list to the new install
-TRANSITIONS=("tor-browser_en-US/Browser/Downloads")
-for dir in ${TRANSITIONS[@]}
-do
-	mkdir -p $TDIR/$dir
-	cp -r --preserve=xattr $TDIR.old/$dir/* $TDIR/$dir
-done
+read -p "Do you want to migrate files to new version? (y/N) " RES
+[ -z "$RES" ] && RES=n
+if [ "$RES" = y ]; then
+	TRANSITIONS=("curr/Browser/Downloads")
+	for dir in ${TRANSITIONS[@]}
+	do
+		mkdir -p $TDIR/$dir
+		cp -r --preserve=xattr $TDIR.old/$dir/* $TDIR/$dir
+	done
+fi
 
 echo "Successfully accompished."
 exit 0
